Add BootstrapServerContext with agent + interlock
